add management command to populate database
- remove populate_database as an api endpoint
- clean up docstrings and type info
